project/requirements.txt
aerich==0.6.1
aiosqlite==0.17.0
anyio==3.6.2
asgiref==3.5.2
asyncpg==0.25.0
attrs==22.1.0
autopep8==2.0.0
click==8.1.3
colorama==0.4.6
ddlparse==1.10.0
dictdiffer==0.9.0
exceptiongroup==1.0.4
fastapi==0.70.1
h11==0.14.0
idna==3.4
iniconfig==1.1.1
iso8601==0.1.16
packaging==22.0
pluggy==1.0.0
pycodestyle==2.10.0
pydantic==1.10.2
pyparsing==3.0.9
pypika-tortoise==0.1.6
pytest==7.2.0
pytz==2022.6
sniffio==1.3.0
starlette==0.16.0
tomli==2.0.1
tomlkit==0.11.6
tortoise-orm==0.17.8
typing_extensions==4.4.0
uvicorn==0.16.0
requests==2.26.0